requirements.txt
tqdm
datasets
tqdm
Pillow
numpy
opencv-python
playwright
requests
matplotlib
duckduckgo_search
langchain
langchain-community
beautifulsoup4
FlagEmbedding
nltk
rouge
rouge_score
peft